Consolidated certain utils.
