+++ src/providers/translationSuggestions/llmCompletion.ts
@@ -7,7 +7,7 @@ import { CodexCellTypes } from "../../../types/enums";
 import { getAutoCompleteStatusBarItem } from "../../extension";
 import { tokenizeText } from "../../utils/nlpUtils";
 import { buildFewShotExamplesText, buildMessages, fetchFewShotExamples, getPrecedingTranslationPairs } from "./shared";
-import { abTestingRegistry } from "../../utils/abTestingRegistry";
+import { abTestingRegistry } from "../../utils/abTesting";
 // A/B testing disabled for now
 
 // Helper function to build A/B test context object
